requirements.txt
flask
Flask-Session
sqlalchemy
psycopg2-binary
python-dotenv
pytesseract
pillow
pypdf
requests
sentence-transformers
pinecone-client
